[web] Add unread indicator to sidebar modals
Summary: This is the second part of [[https://linear.app/comm/issue/ENG-2089/add-unread-indicators-to-sidebar-modal-on-web | ENG-2089]]. This diff adds the basic unread indicators to the sidebar modal. Changes font color and font weight of the last unread message.

Test Plan:
Screenshots that show that unread thread messages appear unread in the sidebar modal in both 'All Threads' and 'My Threads' tabs.

| Unread thread message | Read thread message
| -----  | -----
|{F276779} | {F276782}
| {F276780} {F276781} | {F276783} {F276784}

Reviewers: atul, ginsu, rohan

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D5833