[native] Break spoiler text into individual components for rendering
Summary: Currently, if we wrap spoilers in a `GestureTouchableOpacity`, long spoilers will force a line break both before and after, causing some really bad UI. The purpose of this diff is to handle one of
the potential solutions to this problem, where we tokenize the spoiler text by splitting each word into it's own `Text` component encompassed by a `GestureTouchableOpacity`. This becomes a little complicated
with nested markdown within spoilers, so this is explained by the in-line comments. *This isn't in a state to be landed, but it helps demonstrate the exploration process.*

All context is found in the Linear task: https://linear.app/comm/issue/ENG-2446/[exploration]-break-spoiler-text-into-individual-text-elements

Test Plan: All of the testing videos / screenshots for both iOS and Android can be found in the Linear issue.

Reviewers: ashoat, atul, michal, ginsu

Subscribers: