[desktop] Display a warning when trying to build macos universal
Summary:
More context in [ENG-2214](https://linear.app/comm/issue/ENG-2214). There's an unresolved bug when you are trying to build a universal macos (arm + x64 in one app) version of the app.
It only happens if you already have built the app for one of these versions. This diff adds a warning message that tells you to remove these files if they exist. It also adds a helper script for removing the build files.

Test Plan:
- Build the x64 or arm version
- Try building the universal version
The warning should be displayed
- Remove the files and try building the universal version. It should succeed.

Reviewers: tomek, atul, ginsu, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5869