[lib] Fix fatal error where empty code blocks crash
Summary:
There's a pretty fatal error where an empty code block will
trigger the `throw Error` statement. I suspect that it's because the
object for the message is `{ type: 'codeBlock', content: '\n' }`.

It makes sense to just return the node if we have nothing to traverse within.

https://linear.app/comm/issue/ENG-2438/unexpected-markdown-in-dec-9-daily-update-message-causing-native-app

Test Plan:
Reproduced the message that causes the crash, alongside other more edge cases. Didn't experience a crash
{F278262}
{F278263}

Reviewers: ginsu, atul, tomek, ashoat

Reviewed By: ginsu, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5849