[desktop] Use "os" option for building on windows
Summary:
More discussion in this issue: [ENG-2457](https://linear.app/comm/issue/ENG-2457)
We need to support building `desktop` workspace on Windows for building the electron app for Windows. Unfortunately, we can't just run `yarn install` because it will install dependencies for all workspace members.
The problem with that is that `opaque-ke-napi` and our `postinstall` scripts don't work on Windows and will stop the installation of the rest of the dependencies (e.g. electron).

This diff adds an `os:["!win32"]` option to `package.json`s. By itself this would display an error that "this package is unsupported on this platforms". We can ignore this with a `--ignore-platform` flag (that is now set by default in `.yarnrc`).
The package will still fail installing but crucially the rest of the dependencies will be installed normally.

Additionally in the `native` workspace we set it to `darwin` instead of `!win32` so we can remove the `|| true` safeguard after `pod install` in the `postinstall` script. It was added so the command wouldn't fail on non-macOS environments that don't have `pod`.

Test Plan:
- Test building the desktop app on Windows
- Test building on macOS
- Phab CI passes

Reviewers: tomek, atul, varun

Subscribers: