[Docs] Update nix docs to work with Xcode + iOS
Summary: React Native will run
https://github.com/facebook/react-native/blob/main/scripts/find-node-for-xcode.sh
as part of a post build phase in Xcode, which will search for node in likely
places for macOS.

This enables the developer run the build from Xcode, when following the nix
developer documents.

Currently, the Nix environment would only succeed if ran from the command line
inside of a nix shell.

Related to:
 - https://linear.app/comm/issue/ENG-2318
 - https://linear.app/comm/issue/ENG-2260

Test Plan:

This would be pretty involved, as it requires a clean environment. An already provisioned
development environment will likely have a working setup, and thus wouldn't
see the failures of missing "node".

Steps:
 - Start from clean system
 - Follow setup starting at https://github.com/CommE2E/comm/blob/jonringer/fix-nix-ios-docs-node/docs/nix_dev_env.md#prerequisites and continuing through the mobile prerequisites
 - Launch Xcode
 - Open native/ios workspace
 - Run debug build inside simulator by hitting the "play" button in the upper left

I don't expect people to wipe their systems to test this, more that this would
be rubber stamped.

Reviewers: varun, tomek

Subscribers: