[lib] introduced shimUnsupportedMessageInfo function to reaction message spec
Summary:
introduced `shimUnsupportedMessageInfo` function in reaction message spec. `shimUnsupportedMessageInfo` checks if the user is below the min code version which supports reactions and if that is the case returns an unsupported message info. The robotext for this unsupported message info is "`user` sent a reaction".

Before landing this diff I will go back and change the `minCodeVersion` to be accurate to where we currently are

---

Depends on D5690
Linear Task: [[ https://linear.app/comm/issue/ENG-2245/setup-message-reaction-spec | ENG-2245 ]]

Test Plan:
Please see the screenshots below to see what happens when the reaction message is "unsupported"

What it looks like on the app
{F263891}

The `messageInfo` object
{F263890}

Reviewers: tomek, rohan, atul

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5746