[lib] introduce notificationCollapseKey function to reaction message spec
Summary:
introduce `notificationCollapseKey` function to reaction message spec. For this method, I returned the joined result of the `type`, `threadID`, `creatorID`, and `targetMessageID` fields. By using these fields, a notification will be collapsed and won't alert the user if someone is quickly liking and then unliking the same message. I felt that this was reasonable, but can also see that we might want to remove `targetMessageID` from the joined results to make it less "spammy" if a user likes a bunch of messages in the same thread.

---

Depends on D5870
Linear Task: [[ https://linear.app/comm/issue/ENG-2245/setup-message-reaction-spec | ENG-2245 ]]

Test Plan:
Please watch the demo video to see the changes I made

{F285378}

Reviewers: atul, tomek, rohan, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5871