[lib] Handle string exception in getMessageForException
Summary:
Obvious in hindsight. Only noticed this while debugging an issue where [this line](https://github.com/CommE2E/comm/blob/master/native/input/input-state-container.react.js#L708) was seeing a string exception, which I think is coming from `react-native-background-upload`.

`getMessageForException` is meant to be a general-purpose utility, so I think it's good for us to handle this.

Test Plan: I confirmed that after this change, the `mediaMissionReport` sent by `commbot` included an `exceptionMessage` property that had the text of the exception

Reviewers: tomek

Reviewed By: tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D5890