[native] Fix issue creating Calendar event on Android
Summary:
We've had a long-standing issue with calendar event creation on Android: [ENG-24](https://linear.app/comm/issue/ENG-24/calendar-issues-on-android)

While testing the calendar on Android, I noticed that the background behind `ThreadPickerModal` wasn't showing. I found [this relevant option](https://reactnavigation.org/docs/stack-navigator/#detachpreviousscreen) in the React Navigation docs, which cautioned to instead use `presentation: 'transparentModal'` if you have a transparent modal.

`ThreadPickerModal` and its many siblings are indeed transparent modals, so using `presentation: 'transparentModal'` fixed the issue where the screen behind the modal wasn't showing. Turns out this also solved the bug in calendar event creation, which was happening because the previous screen was unmounted and the complex chain of events that normally handles dismissing the `ThreadPickerModal` after the new `Entry`'s `TextInput` is focused was malfunctioning as a result.

Test Plan: Confirm it's possible to creation a calendar event on Android without the `TextInput` immediately dismissing and the new `Entry` being auto-deleted as a result

Reviewers: inka, tomek

Reviewed By: tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D5891