[native] Fix in-app notif on Android which system theme is dark
Summary:
On Android, we show the same design of in-app notif for both dark mode and light mode. (Unlike iOS, where we vary it.)

However, the default text color on Android depends on the system theme. As a result, if the system theme is dark, then we end up having a white text (default) appear in front of a white background.

This diff makes sure we always set the text color explicitly instead of relying on defaults.

Test Plan: Confirm in-app notif body text can be read

Reviewers: marcin, tomek

Reviewed By: tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D5892