[native] Update comment explaining sleep in Entry on native
Summary:
I remember what this comment was initially referring to. Way back in the past, when we rendered a `TextInput` for every `Entry` regardless of whether it was active or not, we had an issue where the scroll animation for new `Entry` creation was getting stopped partway through when the `TextInput` focused.

This hack appears to no longer be necessary to avoid that issue. However, it turns out removing the hack had a different side-effect, resulting from our new behavior of only rendering a `TextInput` for an active `Entry`. The hack appears to now be necessary for a different reason, which is explained in the updated comment.

Test Plan: Confirm that removing the hack has the effect described in the updated comment, and that removing the hack does not have the effect described in the original comment

Reviewers: tomek

Reviewed By: tomek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D5893