Implement encryption key migration to new accessibility options
Summary: This differential introduces migration for SQLite encryption key we store in secure store in such a way that it is accessible even if the device is locked. expo-secure-store does not provide possibility to update accessibility settings associated with particular key soe it needs to be deleted ans set again with new default options of CommSecureStoreIOSWrapper where new accessbility setting in defined.

Test Plan: Build the app without he changes, set debugger breakpoint in place where you try to access encryption key. Lock the device before you pass throught that breakpoint and wait about a minute. then try to pass through breakpoint. You should not be able to. The add thos changes, build the app a repeate the same steps, but after you launch the app for the first time. Now you should be able to access encryption key.

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: bartek, karol, anunay, ashoat, adrian, abosh

Differential Revision: https://phab.comm.dev/D4795