[services/commtest] Tunnelbroker - Adding of helper function for the string signing with the private key
Summary:
This diff introduces the function to make a signature for the string by the private key. [[ https://www.notion.so/commapp/Tunnelbroker-028bdd578e4c487fb570dbb862952ec6#9758ddb84edd4a9f86ad07a9d8a4b39e | The signature is needed ]] to prove that the public key provided to the Tunnelbroker server belongs to the user during the session creation call.
This function is called during the session creating a call in D5931.

Test Plan: The `commtest` app is successfully built and the tests are successfully passed in the following D5931.

Reviewers: jon, bartek

Reviewed By: jon

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5927