[services/commtest] Tunnelbroker - Adding the tonic client builder function
Summary:
This diff introduces a function to build a Tonic client for further API calls in integrational tests. The pointer to the client will be passed to the helper functions (in D5931) instead of making the "global" client instance.

We are using `COMM_SERVICES_PORT_TUNNELBROKER` environment variable to provide a port to connect to the Tunnelbroker.
The `COMM_SERVICES_HOST_TUNNELBROKER` environment variable provides a host to connect to the Tunnelbroker.

Linear task: [[ https://linear.app/comm/issue/ENG-1657 | ENG-1657 ]]

Test Plan: The `commtest` app is successfully built and the tests are successfully passed in the following D5931 using this function results.

Reviewers: jon, bartek

Reviewed By: jon, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5928