[services/commtest] Tunnelbroker - Adding `deviceID` validation check function
Summary:
This diff introduces a function that makes a call to the Tunnelbroker (get signature API) with the `deviceID` in the wrong format.
In this case, we are making sure that the Tunnelbroker returns an error on the wrong format `deviceID` provided.
This function is called during the test in D5931.

Linear task: [[ https://linear.app/comm/issue/ENG-1657 | ENG-1657 ]]

Test Plan: The `commtest` app is successfully built and the tests are successfully passed in the following D5931 where the function is used.

Reviewers: jon, bartek

Reviewed By: jon

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5929