[services] Tunnelbroker - Adding of the `base64` crate to dependencies
Summary:
This diff adds a `base64` crate to the Cargo Tunnelbroker dependencies.
`base64` functions will be used in the following D5945 in a `verify_signed_string` function.

Linear task: [[ https://linear.app/comm/issue/ENG-2492/removing-cryptopp-from-the-tunnelbroker-and-use-rust-openssl-crate | ENG-2492 ]]

Test Plan:
As this diff is a part of the stack this diff can be tested on the end of the stack at D5947.
To test it:
- Patch to the D5947
- Run the integration tests in D5931 using the `cargo t --test tunnelbroker_integration_test` command.

The expected result is passing the integration test (signing and verifying the string are successful and the new session creation test will pass).

Reviewers: jon, bartek, ashoat

Reviewed By: jon, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5944