[services] Tunnelbroker - Remove `cryptopp` usage from the C++ codebase
Summary:
This diff removes deprecated C++ code related to the use of `cryptopp` library.
In D5947 we are switched to using the Rust function for the signed string verification and we don't need the `cryptopp` and related C++ code base anymore.

Removing `cryptopp` from the dependencies is in the following D5949.

Linear task: [[ https://linear.app/comm/issue/ENG-2492/removing-cryptopp-from-the-tunnelbroker-and-use-rust-openssl-crate | ENG-2492 ]]

Test Plan: The Tunnelbroker server application is successfully built.

Reviewers: jon, bartek, tomek

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5948