[keyserver] modified createMessages function to handle inserting targetMessageID into the DB
Summary:
modified `createdMessages` function to handle inserting `targetMessageID` into the DB. For these changes, I needed to check if a `targetMessageID` existed on the `messageData` object, if it did assign it to the `targetMessageID` variable otherwise assign `targetMessageID` to null. Then we wanted to modify the `messageInsertQuery` to handle setting `targetMessageID` to `target_message` in the db

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2246/liking-a-message | ENG-2246 ]]

Test Plan:
Please see the screenshots below to see what the reaction message looks like in the db. For this example I liked the message "as you inevitably discover bugs, have feature requests..."

{F263940}

Reviewers: atul, rohan, tomek, ashoat

Reviewed By: atul, tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5747