[native/rust_library] Refactor Tonic client function name for Identity service
Summary:
This diff introduces refactoring the Tonic client names for the Identity service in a Rust native library to the `Client` -> `IdentityClient`.
As we are using a single Rust library for the native which will contain clients and helper functions for all services we should refactor the Identity service-related function names to be more specific.
After this, we can add new clients for other services e.g. `TunnelbrokerClient`.

Linear task: [[ https://linear.app/comm/issue/ENG-1333/implement-native-client-for-the-tunnelbroker-api-using-rust | ENG-1333 ]]

Test Plan: Rust library and native app are successfully built.

Reviewers: varun, jon

Reviewed By: varun

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5967