[keyserver] Listen on all addresses in dev mode
Summary:
When we were using Apache as a proxy, we only needed the `keyserver` to listen on `localhost`. That's because Apache was the only thing connected to Node, and Apache would always connect on `localhost`.

Now that clients are connecting directly to Node, we need to tell Node to listen on other addresses. Without this change, it's not possible to connect an external device to a keyserver running on a dev machine.

The reason we didn't catch this in D5846 was because the test plan was insufficient. In the future, we need to be even stricter with test plans.

Test Plan: Make sure a physical iOS device is able to connect to the `keyserver` running on my laptop using port 3000

Reviewers: jon, tomek

Reviewed By: jon

Subscribers: atul

Differential Revision: https://phab.comm.dev/D5973