[web] Only render Loading component in SSR
Summary: This addresses [ENG-2513](https://linear.app/comm/issue/ENG-2513/fix-ssr-on-web). Now that the web app needs to wait on `redux-persist` before hiding `Loading`, all we need to render on the SSR side is `Loading`.

Test Plan: Load the site in both dev and prod modes and make sure there are no warnings about SSR not aligning with client-side rendering, and that there is no "double-render" effect

Reviewers: ginsu, michal, atul, tomek

Reviewed By: ginsu, michal, tomek

Differential Revision: https://phab.comm.dev/D5983