[web] Remove timeZone from Redux state
Summary:
We probably should've never had this in Redux state... would've been more appropriate in React context.

Basically we need to render timestamps in client time instead of server time during SSR, we needed some context to tell components if they needed to shift timestamps to accomodate a timezone.

Following the parent diff, we no longer need to render client timestamps on the server, so we can remove this code.

Test Plan: Flow

Reviewers: ginsu, michal, atul, tomek

Reviewed By: ginsu, michal, tomek

Differential Revision: https://phab.comm.dev/D5984