[bugfix] Fix "unkown error" on login.
Summary:
Fixing this: https://linear.app/comm/issue/ENG-2479/unknown-error-when-password-is-wrong#comment-b1eb427d

We had that problem both on native and web. I introduced new error message in new version,
so it's not confused with anything old anymore. Also updated codeVersion od 167, as I figured we want it ASAP.
I have not set it to 166, as that version was already released.

Code it written in a such way, that code that would run when client has lower version than 167 is identical
to what we had before I introduced changes here: https://phab.comm.dev/rCOMMce7bc1377e206a8fbd8470dfef269556d243728f.

So it is:

```
  if (hasMinCodeVersion(viewer.platformDetails, 167)) {
    // ... new responses
  } else {
    // ... old responses
  }

```

Test Plan:
Tested with codeVersion set to 166 and worked on current native and web.
I had a lot of problems with testing on old clients and I'm not really sure what I've done wrong.
I followed instructions but couldn't make realese build to build and wasted a lot of time on it.
Code is simple enough that it should work on older versions, but it should still be tested. Maybe I'll get some help on that after Christmast in the office.

Reviewers: tomek, kamil, inka, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, atul, inka, kamil, tomek

Differential Revision: https://phab.comm.dev/D5987