[web] Prevent web Menu from dismissing itself on open
Summary:
Some change in React 18 is making `document.addEventListener('click', closeMenuCallback)` execute faster than before, and somehow it's able to catch the click that created it. Not sure how this is happening, but the fix is easy: stop the propagation of the click event once we handle it.

This solves https://linear.app/comm/issue/ENG-2534/cant-press-three-dots-button-in-messagelist-on-web

Test Plan: Make sure I can open the "three dots" menu in the MessageList on web

Reviewers: atul, ginsu, tomek, jon

Reviewed By: atul, ginsu

Differential Revision: https://phab.comm.dev/D6016