[landing] Add `openConnectModal` truthiness check before calling function
Summary:
Still encountering the issue described here: https://linear.app/comm/issue/ENG-2539/encountering-openconnectmodal-is-not-a-function-when-attempting-to

Modifying it to:

{F304010}

got things working again, but not sure if this is a solid solution.

Don't want to get sidetracked by this issue at the moment so haven't really investigated, just wanted to get things back in a working state.

---

Depends on D6026

Test Plan: The error modal no longer appears.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6027