[landing] Don't close SIWEPanel if RainbowKit connect modal successfully completed
Summary: The lines we're updating here are meant to catch RainbowKit closing. The modal will also disappear on success, but in that case `signer` will be set.

Test Plan: Make sure SIWEPanel closes when you press the X in the WebView, but not when you successfully auth with a wallet

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6034