[web] intoduce MessageLikeTooltipButton component
Summary:
introduce `MessageLikeTooltipButton` component. This component returns 👍 wrapped in a div, and the class of the div changes depending on the `viewerReacted` prop that gets passed down. I imagine that this will be a temporary component that will eventually be deleted once we extend from message liking to message reactions

---

Depends on D6055

Linear Task: [[ https://linear.app/comm/issue/ENG-2474/display-that-the-user-has-already-liked-a-message | ENG-2474 ]]

Test Plan:
Please look at the screenshots to see the `MessageLikeTooltipButton` component:

When `viewerReacted` is true:
{F306501}

When `viewerReacted` is false:
{F306500}

Reviewers: atul, tomek

Reviewed By: atul, tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6056