[lib] Introduce `SIWEAuthServerCall` and use as argument for `siweAuth()`
Summary:
Introduce the `SIWEAuthServerCall` type and use as the argument for `siweAuth()`.

We're also updating the return type of the `siwe_auth` endpoint to `boolean` for now. The `boolean` will represent whether or not the `nonce`, `message`, and `signature` in the request are valid and the auth request is "succesful." However, we won't actually register or log in just yet, that will be handled in subsequent diffs (at which point we'll switch from `boolean` to something like `RegisterResponse | LogInResponse`).

---

Depends on D6025

Test Plan: Will be tested implicitly by subsequent diffs... but able to hit endpoint.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6026